requirements.txt
# Core data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Database and data storage
sqlalchemy>=2.0.0
cx_oracle>=8.3.0
oracledb>=1.4.0
python-dotenv>=1.0.0

# Web scraping and API requests
requests>=2.31.0
urllib3>=2.0.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
bokeh>=3.0.0

# Machine learning and statistics
scikit-learn>=1.3.0
statsmodels>=0.14.0
xgboost>=1.7.0
lightgbm>=4.0.0

# Data validation and quality
great-expectations>=0.17.0
pandas-profiling>=3.6.0
datawig>=0.2.0

# Time series analysis
prophet>=1.1.0
statsforecast>=1.5.0

# Geospatial analysis
geopandas>=0.13.0
folium>=0.14.0
shapely>=2.0.0

# Text processing and NLP
nltk>=3.8.0
textblob>=0.17.0

# Data cleaning and preprocessing
pyjanitor>=0.24.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.21.0

# Performance and parallel processing
joblib>=1.3.0
dask>=2023.5.0
numba>=0.57.0

# Configuration and utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
click>=8.1.0

# Jupyter and development
jupyter>=1.0.0
ipywidgets>=8.0.0
jupyter-dash>=0.4.0

# Date and time handling
python-dateutil>=2.8.0
pytz>=2023.3

# API integrations
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data export formats
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Logging and monitoring
structlog>=23.1.0
